requirements.txt
python-gnupg>=0.5.1